services/lti-proxy-databricks/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
httpx==0.27.2
itsdangerous==2.2.0
pydantic==2.9.2
Jinja2==3.1.4
